+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.61.1] - 2020-05-18
+- <no-jira> - Adding viewOnly prop to businessHours redux-form field
+- <no-jira> - Fixing an css issue with break-word property and the MixCalendar and Business Hours Rule components
+
 ## [0.61.0] - 2020-05-12
 - <no-ticket> - Added file upload field component